+++ src/main/GenericBoard.java
@@ -1,6 +1,6 @@
 package main;
 
-import java.awt.Graphics;
+import java.awt.*;
 
 public class GenericBoard extends Board {
 
@@ -14,7 +14,7 @@ public class GenericBoard extends Board {
 	public void paintComponent(Graphics g) {
 		super.paintComponent(g);
 		for (int i = 0; i < cards.size(); i++) {
-			int x = border + (Card.WIDTH + border) * (i % maxCardsX);
+			int x = (int) (border + (Card.WIDTH + border) * (i % maxCardsX));
 			int y = border + (Card.HEIGHT + border) * (int)(i / maxCardsX);
 			if (cards.get(i).isHighlighted()) { y -= border / 2; }
 			cards.get(i).setFaceUp(true);
